… erpnext/accounts/doctype/payment_entry/payment_entry.py
… erpnext/accounts/doctype/pos_invoice/pos_invoice.py
… erpnext/accounts/doctype/purchase_invoice/purchase_invoice.py
… erpnext/accounts/doctype/sales_taxes_and_charges_template/sales_taxes_and_charges_template.py
… erpnext/accounts/party.py
… erpnext/accounts/report/cash_flow/cash_flow.py
… erpnext/accounts/report/consolidated_financial_statement/consolidated_financial_statement.py
… erpnext/accounts/report/customer_ledger_summary/customer_ledger_summary.py
… erpnext/accounts/report/dimension_wise_accounts_balance_report/dimension_wise_accounts_balance_report.py
… erpnext/accounts/report/financial_statements.py
… erpnext/accounts/report/general_ledger/general_ledger.py
… erpnext/accounts/report/trial_balance/trial_balance.py
… erpnext/accounts/utils.py