Similar to #32776 

Applied on **Accounts** Module

### Before

```
In [1]: from erpnext.accounts.doctype.fiscal_year.fiscal_year import get_from_and_to_date
In [2]: %timeit get_from_and_to_date('2022-2023')
2.52 ms ± 63.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

### After
```
In [1]: from erpnext.accounts.doctype.fiscal_year.fiscal_year import get_from_and_to_date
In [2]: %timeit get_from_and_to_date('2022-2023')
5.84 µs ± 203 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```